create_package: don't delete linux scripts for armv6 & 7

+++ scripts/dev/create_package.sh
@@ -357,7 +357,7 @@ function createPackage {
 
     #delete scripts
     cd $pkg_ofroot/scripts
-	if [ "$pkg_platform" != "linux64" ]; then
+	if [ "$pkg_platform" != "linux64" ] && [ "$pkg_platform" != "linuxarmv6l" ] && [ "$pkg_platform" != "linuxarmv7l" ]; then
     	rm -Rf $otherplatforms
 	else
     	rm -Rf win_cb vs osx ios
